fix flats

+++ house24/admin_panel/views/flat_views.py
@@ -22,27 +22,37 @@ class PostInstanceMixin:
     def save_form(self, form, request):
         if form.is_valid():
             obj = form.save()
-            personal_account = PersonalAccount.find_inst(form.cleaned_data['account'])
-            if personal_account:
-                # if this account already exists  - set it as account for this flat
-                obj.personal_account = personal_account
-                obj.save()
-                return redirect('admin_panel:list_flats_admin')
-            else:
-                #  otherwise - we will create new personal account using given data
-                data = {'number': form.cleaned_data['account'],
-                        'house': form.cleaned_data['house'],
-                        'section': form.cleaned_data['section']}
-                account_form = PersonalAccountForm(data)
-                if account_form.is_valid():
-                    obj.personal_account = account_form.save()
-                    obj.save()
+            account_data = form.cleaned_data.get('account')
+            if account_data:
+                personal_account = PersonalAccount.find_inst(account_data)
+                if personal_account:
+                    # if this account already exists  - set it as account for this flat
+                    old_account = obj.account
+                    old_account.flat = None  # clear old one-to-one field
+                    old_account.save()
+                    personal_account.flat = obj
+                    personal_account.save()
                     return redirect('admin_panel:list_flats_admin')
                 else:
-                    obj.delete()
-
+                    #  otherwise - we will create new personal account using given data
+                    data = {'number': form.cleaned_data['account'],
+                            'house': form.cleaned_data.get('house'),
+                            'section': form.cleaned_data.get('section')}
+                    account_form = PersonalAccountForm(data)
+                    if account_form.is_valid():
+                        new_account = account_form.save()
+                        new_account.flat = obj
+                        new_account.save()
+                        return redirect('admin_panel:list_flats_admin')
+                    else:
+                        obj.delete()
+                        personal_accounts = PersonalAccount.objects.filter(flat__isnull=True)
+                        return render(request, self.template_name, context={'form': form,
+                                                                            'personal_accounts': personal_accounts})
+            else:
+                return redirect('admin_panel:list_flats_admin')
         else:
-            personal_accounts = PersonalAccount.objects.filter(flats__isnull=True)
+            personal_accounts = PersonalAccount.objects.filter(flat__isnull=True)
             return render(request, self.template_name, context={'form': form,
                                                                 'personal_accounts': personal_accounts})
 
@@ -55,7 +65,7 @@ class CreateFlatView(AdminPermissionMixin, View, PostInstanceMixin):
 
     def get(self, request, pk=None):
         form = self.form_class()
-        personal_accounts = PersonalAccount.objects.filter(flats__isnull=True)
+        personal_accounts = PersonalAccount.objects.filter(flat__isnull=True)
         return render(request, self.template_name, context={'form': form,
                                                             'personal_accounts': personal_accounts})
 
@@ -73,11 +83,11 @@ class UpdateFlatView(AdminPermissionMixin, View, PostInstanceMixin):
     def get(self, request, pk):
         obj = get_object_or_404(self.model, pk=pk)
         try:
-            personal_account = obj.personal_account.number
+            personal_account = obj.account.number
         except AttributeError:
             personal_account = None
         form = CreateFlatForm(instance=obj, **{'account_number': personal_account})
-        personal_accounts = PersonalAccount.objects.filter(flats__isnull=True)
+        personal_accounts = PersonalAccount.objects.filter(flat__isnull=True)
         return render(request, self.template_name, context={'form': form,
                                                             'personal_accounts': personal_accounts})
